update getFileSystemAndRelativePathByPath and test

+++ lib/src/filesystem/disk.dart
@@ -8,9 +8,17 @@ class Disk implements IFileSystem {
 
   List<String> get targets => _fileSystems.keys.toList();
 
-  (IFileSystem?, String) getFileSystemAndRelativePathByPath(String path) {
-    final target = rootPrefix(path);
-    return (_fileSystems[target], relative(path, from: target));
+  (IFileSystem?, String?) getFileSystemAndRelativePathByPath(String path) {
+    var target = rootPrefix(path);
+    if (target != "") {
+      var relativePath = relative(path, from: target);
+      if (relativePath == ".") {
+        return (_fileSystems[target], r"\");
+      } else {
+        return (_fileSystems[target], "\\$relativePath");
+      }
+    }
+    return (null, null);
   }
 
   void mount(String target, IFileSystem fs) {
@@ -44,7 +52,7 @@ class Disk implements IFileSystem {
   @override
   Entity? open(String path) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(path);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       if (fs.open(rPath) case Entity entity) {
         entity.path = join(rootPrefix(path), entity.path);
         return entity;
@@ -60,7 +68,7 @@ class Disk implements IFileSystem {
     EntityType type = EntityType.file,
   }) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(path);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       return fs.create(rPath, recursive: recursive, type: type);
     }
     throw "[Disk]: FileSystems don't mount.";
@@ -69,7 +77,7 @@ class Disk implements IFileSystem {
   @override
   Entity copy(String filePath, String dirPath) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(filePath);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       return fs.copy(rPath, dirPath);
     }
     throw "[Disk]: FileSystems don't mount.";
@@ -78,7 +86,7 @@ class Disk implements IFileSystem {
   @override
   Entity move(String pathIn, String pathOut) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(pathIn);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       return fs.copy(rPath, pathOut);
     }
     throw "[Disk]: FileSystems don't mount.";
@@ -87,7 +95,7 @@ class Disk implements IFileSystem {
   @override
   bool delete(String path, {bool recursive = false}) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(path);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       return fs.delete(rPath, recursive: recursive);
     }
     throw "[Disk]: FileSystems don't mount.";
@@ -96,7 +104,7 @@ class Disk implements IFileSystem {
   @override
   Stream<Entity> list(String dirPath) async* {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(dirPath);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       yield* fs.list(rPath);
     }
   }
@@ -104,7 +112,7 @@ class Disk implements IFileSystem {
   @override
   Stream<int> openRead(String filePath) async* {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(filePath);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       yield* fs.openRead(rPath);
     }
   }
@@ -112,7 +120,7 @@ class Disk implements IFileSystem {
   @override
   Stat stat(String path) {
     var (fs, rPath) = getFileSystemAndRelativePathByPath(path);
-    if (fs != null) {
+    if (fs != null && rPath != null) {
       return fs.stat(rPath);
     }
     throw "[Disk]: FileSystems don't mount.";
